refactor(frontend): seperate Alerts into own component
