+++ src/frontend/src/components/nav/Header.tsx
@@ -1,20 +1,14 @@
 import {
   ActionIcon,
-  Alert,
   Container,
   Group,
   Indicator,
-  Menu,
   Tabs,
   Text,
   Tooltip
 } from '@mantine/core';
 import { useDisclosure } from '@mantine/hooks';
-import {
-  IconBell,
-  IconExclamationCircle,
-  IconSearch
-} from '@tabler/icons-react';
+import { IconBell, IconSearch } from '@tabler/icons-react';
 import { useQuery } from '@tanstack/react-query';
 import { type ReactNode, useEffect, useMemo, useState } from 'react';
 import { useMatch, useNavigate } from 'react-router-dom';
@@ -34,21 +28,14 @@ import {
 import { useUserState } from '../../states/UserState';
 import { ScanButton } from '../buttons/ScanButton';
 import { SpotlightButton } from '../buttons/SpotlightButton';
+import { Alerts } from './Alerts';
 import { MainMenu } from './MainMenu';
 import { NavHoverMenu } from './NavHoverMenu';
 import { NavigationDrawer } from './NavigationDrawer';
 import { NotificationDrawer } from './NotificationDrawer';
 import { SearchDrawer } from './SearchDrawer';
 
-interface AlertInfo {
-  key: string;
-  title: string;
-  message: string;
-}
-
 export function Header() {
-  const user = useUserState();
-
   const [setNavigationOpen, navigationOpen] = useLocalState((state) => [
     state.setNavigationOpen,
     state.navigationOpen
@@ -73,49 +60,6 @@ export function Header() {
   const navbar_message = useMemo(() => {
     return server.customize?.navbar_message;
   }, [server.customize]);
-
-  const [dismissed, setDismissed] = useState<string[]>([]);
-
-  const alerts: AlertInfo[] = useMemo(() => {
-    const _alerts: AlertInfo[] = [];
-
-    if (server?.debug_mode) {
-      _alerts.push({
-        key: 'debug',
-        title: t`Debug Mode`,
-        message: t`The server is running in debug mode.`
-      });
-    }
-
-    if (server?.worker_running == false) {
-      _alerts.push({
-        key: 'worker',
-        title: t`Background Worker`,
-        message: t`The background worker process is not running.`
-      });
-    }
-
-    if (globalSettings.isSet('SERVER_RESTART_REQUIRED')) {
-      _alerts.push({
-        key: 'restart',
-        title: t`Server Restart`,
-        message: t`The server requires a restart to apply changes.`
-      });
-    }
-
-    const n_migrations =
-      Number.parseInt(globalSettings.getSetting('_PENDING_MIGRATIONS')) ?? 0;
-
-    if (n_migrations > 0) {
-      _alerts.push({
-        key: 'migrations',
-        title: t`Database Migrations`,
-        message: t`There are pending database migrations.`
-      });
-    }
-
-    return _alerts.filter((alert) => !dismissed.includes(alert.key));
-  }, [server, dismissed, globalSettings]);
 
   // Fetch number of notifications for the current user
   const notifications = useQuery({
@@ -213,35 +157,7 @@ export function Header() {
                 </ActionIcon>
               </Tooltip>
             </Indicator>
-            {user.isStaff() && alerts.length > 0 && (
-              <Menu withinPortal={true} position='bottom-end'>
-                <Menu.Target>
-                  <Tooltip position='bottom-end' label={t`Alerts`}>
-                    <ActionIcon
-                      variant='transparent'
-                      aria-label='open-alerts'
-                      color='red'
-                    >
-                      <IconExclamationCircle />
-                    </ActionIcon>
-                  </Tooltip>
-                </Menu.Target>
-                <Menu.Dropdown>
-                  {alerts.map((alert) => (
-                    <Menu.Item key={`alert-item-${alert.key}`}>
-                      <Alert
-                        withCloseButton
-                        color='red'
-                        title={alert.title}
-                        onClose={() => setDismissed([...dismissed, alert.key])}
-                      >
-                        {alert.message}
-                      </Alert>
-                    </Menu.Item>
-                  ))}
-                </Menu.Dropdown>
-              </Menu>
-            )}
+            <Alerts />
             <MainMenu />
           </Group>
         </Group>
